Separate functions for OAS 3.1 from OAS 3.0 functions
